web: replace SocketService with useSocket hook
also refactor socket page into a server component
TODO: need to reimplement game ID, user ID validation logic inside the
server component
