+++ web/package.json
@@ -36,6 +36,7 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.3.1",
+    "@next/eslint-plugin-next": "^15.3.1",
     "@tailwindcss/postcss": "^4.1.4",
     "@types/node": "^20.17.30",
     "@types/pg": "^8.11.13",
@@ -43,6 +44,7 @@
     "@types/react-dom": "^19.1.2",
     "eslint": "^9.24.0",
     "eslint-config-next": "15.2.4",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "loader-utils": "^1.4.2",
     "prettier": "^2.8.8",
     "prisma": "^6.6.0",
